+++ lib/app/features/data_subject_right/screens/create_dsr/accepte_term_screen.dart
